refactor: 카드 게임 상태 데이터 타입 수정 및 불필요한 카드 상수 제거
CardType 같은 타입이 2번 중복으로 정의되어 있어, 하나로 통합함
