+++ frontend/src/contexts/CardGame/CardGameProvider.tsx
@@ -1,10 +1,17 @@
 import { useWebSocketSubscription } from '@/apis/websocket/hooks/useWebSocketSubscription';
 import { CardGameRound } from '@/constants/miniGame';
-import { CardGameState, CardGameStateData, CardInfo, SelectedCardInfo } from '@/types/miniGame';
+import { CardGameState, CardInfo, SelectedCardInfo } from '@/types/miniGame';
 import { PropsWithChildren, useCallback, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useIdentifier } from '../Identifier/IdentifierContext';
 import { CardGameContext } from './CardGameContext';
+
+type CardGameStateResponse = {
+  cardGameState: CardGameState;
+  currentRound: CardGameRound;
+  cardInfoMessages: CardInfo[];
+  allSelected: boolean;
+};
 
 const CardGameProvider = ({ children }: PropsWithChildren) => {
   const navigate = useNavigate();
@@ -28,7 +35,7 @@ const CardGameProvider = ({ children }: PropsWithChildren) => {
   });
 
   const handleCardGameState = useCallback(
-    (data: CardGameStateData) => {
+    (data: CardGameStateResponse) => {
       const { cardGameState, currentRound, cardInfoMessages } = data;
 
       const isPreparing = cardGameState === 'PREPARE';
